save_answers arguments changed from float to double

+++ include/ISearchEngine.h
@@ -19,7 +19,7 @@ public:
     virtual ~ISearchEngine() = default;
     virtual std::vector<std::vector<RelativeIndex>> search(const std::vector<std::string>& queries_input) = 0;
     virtual std::vector<RelativeIndex> search(std::string query) = 0;
-    virtual void set_response_limit(size_t limit) = 0;
+    virtual void set_response_limit(int limit) = 0;
 
 };
