+++ include/SearchEngine.h
@@ -11,11 +11,11 @@ class SearchEngine : public ISearchEngine
 private:
     std::vector<std::unordered_set<std::string>> _query_set;
     FreqDict _freq_dict;
-    size_t _response_limit = 5;
+    int _response_limit = 5;
 
 public:
     SearchEngine(IInvertedIndex& idx);
     std::vector<std::vector<RelativeIndex>> search(const std::vector<std::string>& queries_input) override;
     std::vector<RelativeIndex> search(std::string query) override;
-    void set_response_limit(size_t limit) override;
+    void set_response_limit(int limit) override;
 };
